check if there are any instances before delete

+++ run.sh
@@ -30,8 +30,12 @@ function cleanup() {
 
   clusters=$(gcloud container clusters list --format json | jq -r '.[].name | select(. | contains("icluster-kube-"))')
   instances=$(gcloud compute instances list --filter "tags:docker-machine" --format json | jq -r '.[].name')
-  gcloud container clusters delete -q $clusters
-  gcloud compute instances delete -q $instances
+  if [ ! -z "$clusters" ]; then
+    gcloud container clusters delete -q $clusters
+  fi
+  if [ ! -z "$instances" ]; then
+    gcloud compute instances delete --delete-disks -q $instances
+  fi
 }
 
 if which apt-get; then
